feat(toolkit): hide amazon q explorer tree when its installed
- Also remove a lot of now irrelevant code, and move some code to more appropriate files.
